requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
black==23.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.90.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.23.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
